requirements.txt
Python 3.13.2
Flask 3.1.0
psycopg2-binary 2.9.10
SQLAlchemy        2.0.38